exercise 4.21

+++ part4/controllers/blogs.js
@@ -18,6 +18,7 @@ const getTokenFrom = (request) => {
 
 blogsRouter.post("/", async (request, response) => {
   const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  console.log(decodedToken);
   if (!decodedToken.id) {
     return response.status(401).json({ error: "token missing or invalid" });
   }
@@ -41,8 +42,21 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id);
-  response.status(204).end();
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
+
+  const blog = await Blog.findById(request.params.id);
+
+  if (blog.user.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id);
+    response.status(204).end();
+  } else {
+    return response
+      .status(401)
+      .json({ error: "Cannot delete someone else's blog." });
+  }
 });
 
 blogsRouter.put("/:id", async (request, response) => {
